Adding coderabbit fix suggestions (bugged)

+++ server/src/main/java/com/nestorria/server/modules/booking/dto/AgencyDashboardResponse.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 public record AgencyDashboardResponse(
     int totalBookings,
-    int totalRevenue,
+    long totalRevenue,
     List<BookingResponse> bookings
 ) {}
